server stress test script support
