+++ server/package.json
@@ -86,7 +86,8 @@
     "supertest": "6.2.3",
     "ts-jest": "^28.0",
     "ts-node": "^10.7.0",
-    "typescript": "4.6.3"
+    "typescript": "4.6.3",
+    "wrtc": "^0.4.7"
   },
   "resolutions": {
     "bl": "4.0.3"
