+++ server/src/config.ts
@@ -54,6 +54,12 @@ const SPACES_ENDPOINT =
 
 const JWTSECRET = process.env.JWTSECRET;
 
+const CALLBACK_DEBOUNCE_WAIT = parseInt(process.env.CALLBACK_DEBOUNCE_WAIT) || 2000;
+const CALLBACK_DEBOUNCE_MAXWAIT = parseInt(process.env.CALLBACK_DEBOUNCE_MAXWAIT) || 10000;
+const CALLBACK_URL = process.env.CALLBACK_URL ? new URL(process.env.CALLBACK_URL) : null;
+const CALLBACK_TIMEOUT = parseInt(process.env.CALLBACK_TIMEOUT) || 5000;
+const CALLBACK_OBJECTS = process.env.CALLBACK_OBJECTS ? JSON.parse(process.env.CALLBACK_OBJECTS) : {};
+
 const DEFAULT_RELM_CONTENT = {
   entities: [
     {
@@ -149,5 +155,10 @@ export {
   FACEBOOK_CLIENT_ID,
   FACEBOOK_CLIENT_SECRET,
   TWITTER_CLIENT_ID,
-  TWITTER_CLIENT_SECRET
+  TWITTER_CLIENT_SECRET,
+  CALLBACK_DEBOUNCE_WAIT,
+  CALLBACK_DEBOUNCE_MAXWAIT,
+  CALLBACK_URL,
+  CALLBACK_OBJECTS,
+  CALLBACK_TIMEOUT
 };
