+++ server/src/scripts/client/RelmParticipant.ts
@@ -1,12 +1,11 @@
-import * as awarenessProtocol from "y-protocols/awareness";
-import * as time from "lib0/time";
-import * as f from "lib0/function";
+import { RelmProvider } from "relm-common";
+import { LocalState } from "./LocalState.js";
 
 export type ParticipantOptions = {
   clientId: number,
   participantId: string,
   name: string,
-  awareness: awarenessProtocol.Awareness
+  provider: RelmProvider
 };
 
 const HEX_CHARS = "01234567890ABCDEF".split("");
@@ -23,9 +22,13 @@ function randomHexChar() {
 export class RelmParticipant {
 
   private options: ParticipantOptions;
+  private tcpState: LocalState;
+  private udpState: LocalState;
 
   constructor(options: ParticipantOptions) {
     this.options = options;
+    this.tcpState = new LocalState(options.clientId, options.provider.ws.awareness);
+    this.udpState = new LocalState(options.clientId, options.provider.gecko.awareness);
   }
 
   join() {
@@ -33,82 +36,29 @@ export class RelmParticipant {
     const [spawnX, spawnY, spawnZ] = this.getSpawnPos();
 
     // Broadcast our spawning to other players
-    this.options.awareness.states.set(this.options.clientId, {});
-    this.setLocalField("id", this.options.participantId);
-    this.setLocalField("a", { clipIndex: 2, animLoop: true });
-    this.setLocalField("user", { color: randomColor(), name: this.options.name });
-    this.setLocalField("i", this.getBotIdentityData());
+    this.options.provider.ws.awareness.states.set(this.options.clientId, {});
+    this.options.provider.gecko.awareness.states.set(this.options.clientId, {});
+
+    this.tcpState.setLocalField("id", this.options.participantId);
+    this.tcpState.setLocalField("user", { color: randomColor(), name: this.options.name });
+    this.tcpState.setLocalField("i", this.getBotIdentityData());
+
+    this.udpState.setLocalField("id", this.options.participantId);
+    this.udpState.setLocalField("a", { clipIndex: 2, animLoop: true });
+
     this.setTransform([spawnX, spawnY, spawnZ, 0, 0, 0]);
   }
 
   getSpawnPos() {
-    return (this.options.awareness.doc.getMap("entryways").get("default") || [0, 0, 0]) as number[];
+    return (this.options.provider.ws.awareness.doc.getMap("entryways").get("default") || [0, 0, 0]) as number[];
   }
 
   getTransform() {
-    return this.getLocalField("t");
+    return this.udpState.getLocalField("t");
   }
 
   setTransform(transform: number[]) {
-    this.setLocalField("t", transform);
-  }
-
-  getLocalState() {
-    return this.options.awareness.states.get(this.options.clientId) || null;
-  }
-
-  setLocalState(state: any) {
-    const clientID = this.options.clientId;
-    const currLocalMeta = this.options.awareness.meta.get(clientID);
-    const clock = currLocalMeta === undefined ? 0 : currLocalMeta.clock + 1;
-    const prevState = this.options.awareness.states.get(clientID);
-
-    if (state === null) {
-      this.options.awareness.states.delete(clientID);
-    } else {
-      this.options.awareness.states.set(clientID, state);
-    }
-    this.options.awareness.meta.set(clientID, {
-      clock,
-      lastUpdated: time.getUnixTime()
-    });
-    const added = [];
-    const updated = [];
-    const filteredUpdated = [];
-    const removed = [];
-    if (state === null) {
-      removed.push(clientID);
-    } else if (prevState == null) {
-      if (state != null) {
-        added.push(clientID);
-      }
-    } else {
-      updated.push(clientID);
-      if (!f.equalityDeep(prevState, state)) {
-        filteredUpdated.push(clientID);
-      }
-    }
-    if (added.length > 0 || filteredUpdated.length > 0 || removed.length > 0) {
-      this.options.awareness.emit('change', [{ added, updated: filteredUpdated, removed }, 'local']);
-    }
-    this.options.awareness.emit('update', [{ added, updated, removed }, 'local']);
-  }
-
-  getLocalField(field: string) {
-    const state = this.getLocalState();
-    if (state !== null) {
-      return state[field];
-    }
-  }
-
-  setLocalField(field: string, value: any) {
-    const state = this.getLocalState();
-    if (state !== null) {
-      this.setLocalState({
-        ...state,
-        [field]: value
-      });
-    }
+    this.udpState.setLocalField("t", transform);
   }
 
   private getBotIdentityData() {
